Updated algorithms struct + Makefiles
