+++ algorithms/cc/Makefile.include
@@ -1,0 +1,13 @@
+# Directories
+CC_DIR := algorithms/cc
+
+# Sources
+CC_SRC := $(CC_DIR)/cc.cpp
+ALG_SRC += $(CC_SRC)
+
+# Objects
+ALG_OBJ += $(addprefix $(BUILD_DIR)/, $(CC_SRC:.cpp=.o))
+
+# Build folders
+create-folders::
+	@mkdir -p $(BUILD_DIR)/$(CC_DIR)
