+++ algorithms/cc/cc.hpp
@@ -1,8 +1,6 @@
-/** @file defs.hpp
+/** @file cc.hpp
 
- @brief <b>Common definitions for all SBG modules</b>
- 
- This file provides helpful definitions that will be used by all SBG modules.
+ @brief <b>Connected Components SBG implementation</b>
 
  <hr>
 
@@ -23,39 +21,22 @@
 
  ******************************************************************************/
 
-#ifndef SBG_DEFS_HPP
-#define SBG_DEFS_HPP
+#ifndef SBG_CC_HPP
+#define SBG_CC_HPP
+
+#include "sbg/sbg.hpp"
 
 namespace SBG {
 
-namespace Util {
+namespace LIB {
 
 ////////////////////////////////////////////////////////////////////////////////
-// Helpful macros --------------------------------------------------------------
+// Connected components --------------------------------------------------------
 ////////////////////////////////////////////////////////////////////////////////
 
-/**
- * @brief Define a variable member of a class and the getters and setters.
- *
- * The macro takes two parameters:
- *  + The variable type @c X.
- *  + The name of the variable @c Y.
- *
- * It will define the following:
- *  + A member variable called: @c Y_
- *  + A constant ref getter method: @c Y()
- *
- */
-#define member_class(X, Y) \
-  private:                 \
-  X Y##_;                  \
-  public:                  \
-  const X &Y() const;
+PWMap connectedComponents(SBG g);
 
-#define member_imp(C, X, Y)          \
-  const X &C::Y() const { return Y##_; }
-
-} // namespace Util
+} // namespace LIB
 
 } // namespace SBG
