+++ algorithms/cutvertex/Makefile.include
@@ -1,0 +1,13 @@
+# Directories
+CV_DIR := algorithms/cutvertex
+
+# Sources
+CV_SRC := $(CV_DIR)/cut_vertex.cpp
+ALG_SRC += $(CV_SRC)
+
+# Objects
+ALG_OBJ += $(addprefix $(BUILD_DIR)/, $(CV_SRC:.cpp=.o))
+
+# Build folders
+create-folders::
+	@mkdir -p $(BUILD_DIR)/$(CV_DIR)
