+++ algorithms/misc/Makefile.include
@@ -1,0 +1,14 @@
+# Directories
+MISC_DIR := algorithms/misc
+
+# Sources
+MISC_SRC := $(MISC_DIR)/causalization_builders.cpp \
+  $(MISC_DIR)/causalization_json.cpp
+ALG_SRC += $(MISC_SRC)
+
+# Objects
+MISC_OBJ = $(addprefix $(BUILD_DIR)/, $(MISC_SRC:.cpp=.o))
+
+# Build folders
+create-folders::
+	@mkdir -p $(BUILD_DIR)/$(MISC_DIR)
