+++ algorithms/scc/Makefile.include
@@ -1,0 +1,13 @@
+# Directories
+SCC_DIR := algorithms/scc
+
+# Sources
+SCC_SRC := $(SCC_DIR)/scc.cpp
+ALG_SRC += $(SCC_SRC)
+
+# Objects
+ALG_OBJ += $(addprefix $(BUILD_DIR)/, $(SCC_SRC:.cpp=.o))
+
+# Build folders
+create-folders::
+	@mkdir -p $(BUILD_DIR)/$(SCC_DIR)
