+++ algorithms/scc/scc.hpp
@@ -1,8 +1,6 @@
-/** @file defs.hpp
+/** @file scc.hpp
 
- @brief <b>Common definitions for all SBG modules</b>
- 
- This file provides helpful definitions that will be used by all SBG modules.
+ @brief <b>SCC SBG implementation</b>
 
  <hr>
 
@@ -23,39 +21,56 @@
 
  ******************************************************************************/
 
-#ifndef SBG_DEFS_HPP
-#define SBG_DEFS_HPP
+#ifndef SBG_SCC_HPP
+#define SBG_SCC_HPP
+
+#include "sbg/sbg.hpp"
 
 namespace SBG {
 
-namespace Util {
+namespace LIB {
 
 ////////////////////////////////////////////////////////////////////////////////
-// Helpful macros --------------------------------------------------------------
+// SCC -------------------------------------------------------------------------
 ////////////////////////////////////////////////////////////////////////////////
 
-/**
- * @brief Define a variable member of a class and the getters and setters.
- *
- * The macro takes two parameters:
- *  + The variable type @c X.
- *  + The name of the variable @c Y.
- *
- * It will define the following:
- *  + A member variable called: @c Y_
- *  + A constant ref getter method: @c Y()
- *
- */
-#define member_class(X, Y) \
-  private:                 \
-  X Y##_;                  \
-  public:                  \
-  const X &Y() const;
+struct SCC {
+  private:
+  const PWMapAF &fact_;
 
-#define member_imp(C, X, Y)          \
-  const X &C::Y() const { return Y##_; }
+  //*** SBG info, constant
+  member_class(DSBG, dsbg);
 
-} // namespace Util
+  member_class(Set, V);
+  member_class(PWMap, Vmap);
+
+  member_class(PWMap, Emap);
+  member_class(PWMap, subEmap);
+
+  //-----------------------------
+  member_class(Set, E); // Edges in the same SCC in each step
+  member_class(Set, Ediff); // Edges between different SCC in each step
+
+  member_class(PWMap, mapB);
+  member_class(PWMap, mapD);
+ 
+  member_class(PWMap, rmap);
+
+  member_class(bool, debug);
+
+  public:
+  SCC(const DSBG &dsbg, bool debug);
+
+  PWMap calculate();
+
+  const PWMapAF &fact() const;
+
+  private:
+  PWMap sccMinReach(const DSBG &dg) const;
+  PWMap sccStep();
+};
+
+} // namespace LIB
 
 } // namespace SBG
