+++ eval/defs.hpp
@@ -32,7 +32,13 @@
 #include "ast/expr.hpp"
 #include "ast/statement.hpp"
 #include "sbg/af_pwmap.hpp"
-#include "sbg/sbg_algorithms.hpp"
+#include "algorithms/cc/cc.hpp"
+#include "algorithms/matching/matching.hpp"
+#include "algorithms/scc/scc.hpp"
+#include "algorithms/cutvertex/cut_vertex.hpp"
+#include "algorithms/toposort/topo_sort.hpp"
+#include "algorithms/misc/causalization_builders.hpp"
+#include "algorithms/misc/causalization_json.hpp"
 
 namespace SBG {
